fixes #5114

This is a really big change. I think it reduced the build time, but can't say for sure due to processes running on my computer. Here is a really small sample of `yarn compile` I did in the master branch.

```
✨  Done in 35.85s.
✨  Done in 35.51s.
✨  Done in 35.07s.
✨  Done in 34.73s.
```

This is a sample of build times on this branch.

```
✨  Done in 35.98s.
✨  Done in 35.37s.
✨  Done in 35.34s.
✨  Done in 34.21s.
```

My localized performance test said 100,000 iterations of `setter` vs `if` statement is ~900ms slower. I'm not sure if this is worth merging. I'm going to mark this as draft so we can discuss.