@matuzalemsteles 2022-11-30
I thought I commented on this, maybe I just commented about this on Slack.

I'm not sure if that could actually be a huge performance improvement, although `setter` is slower it could be that it's just noticeable in a very large amount as you did for 100,000 interactions, perhaps for our codebase this is not noticeable because we have a much smaller amount.

Even though you have other things running on your machine, this is validate the numbers because you compare the two changes in the same environment so this shouldn't be a problem.