@pat270 2023-02-06
Closing this one, we can come back to it later if we need to fine tune performance.